+++ integration/gotests/gotest_test.go
@@ -55,6 +55,7 @@ func testPkgs(t *testing.T) []string {
 		// These have special configuration.
 		"github.com/u-root/u-root/pkg/gpio",
 		"github.com/u-root/u-root/pkg/mount",
+		"github.com/u-root/u-root/pkg/mount/block",
 		"github.com/u-root/u-root/pkg/mount/loop",
 
 		// Missing xzcat in VM.
